Fixed xcode 9 compiler warnings

+++ include/taco/format.h
@@ -32,7 +32,7 @@ public:
   /// position i is specified by modeTypes[i]. The mode stored in position i is
   /// specified by modeOrdering[i].
   Format(const std::vector<ModeType>& modeTypes,
-         const std::vector<int>& modeOrdering);
+         const std::vector<size_t>& modeOrdering);
 
   /// Returns the number of modes in the format.
   size_t getOrder() const;
@@ -43,11 +43,11 @@ public:
 
   /// Get the ordering in which the modes are stored. The mode stored in
   /// position i is specifed by element i of the returned vector.
-  const std::vector<int>& getModeOrdering() const;
+  const std::vector<size_t>& getModeOrdering() const;
 
 private:
   std::vector<ModeType> modeTypes;
-  std::vector<int>      modeOrdering;
+  std::vector<size_t>   modeOrdering;
 };
 
 bool operator==(const Format&, const Format&);
